reload SpendingList on adding a spending
SpendingsPage will coordinate 3 pieces of state:

- the current order
- the current filter
- the fact that the form just successfully created an element

SpendingPage passes `newestSpending` to SpendingList because it doesn't
care how SpendingList handles the new data. SpendingList may want to go
to the API or may just want to present the newest spending without a
roundtrip.

(I'm planning on implementing ordering and filterng on the back-end, so
going to the API will in fact be SpendingList's choice.)